requirements.txt
appnope==0.1.0
attrs==19.1.0
awscli==1.16.250
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blis==0.4.1
boto==2.49.0
boto3==1.9.240
botocore==1.12.240
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
cssselect==1.1.0
cvxpy==1.0.25
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
ecos==2.0.7.post1
en-core-web-md==2.2.0
en-core-web-sm==2.2.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
findspark==1.3.0
flashtext==2.7
future==0.18.2
fuzzywuzzy==0.18.0
gensim==3.8.1
goose3==3.1.6
gzip-stream==1.0.0
idna==2.8
ipykernel==5.1.2
ipysheet==0.4.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
jieba==0.39
jieba3k==0.35.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
multiprocess==0.70.9
multitasking==0.0.9
murmurhash==1.0.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
newspaper3k==0.2.8
nltk==3.4.5
notebook==6.0.1
numpy==1.17.2
osqp==0.6.1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pke==1.8.1
plac==0.9.6
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.11.0
ptyprocess==0.6.0
py4j==0.10.7
pyasn1==0.4.7
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pyspark==2.4.4
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.2
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
requests-file==1.4.3
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
scs==2.1.1.post2
selectolax==0.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.4
spacy==2.2.0
srsly==0.1.0
stanfordnlp==0.2.0
statsmodels==0.10.2
terminado==0.8.2
testpath==0.4.2
thinc==7.1.1
tinysegmenter==0.3
tldextract==2.2.1
torch==1.3.1
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
unicodedata2==12.0.0
Unidecode==1.1.1
urllib3==1.25.6
warc==0.2.1
warc3-wet==0.2.3
warcio==1.7.1
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
yfinance==0.1.52
